requirements.txt
fastai==1.0.61
librosa
pyaudio
python-pushover
